requirements.txt
python-dotenv
requests
spotipy
pandas
numpy
scikit-learn
DateTime
ipykernel